requirements-dev.txt
# Development dependencies (pinned for reproducible installs)
pytest-cov==4.0.0
pytest-subtests==0.11.0
black==24.4.2
ruff==0.5.6
mypy==1.10.0
types-PyYAML==6.0.12.20240311
pre-commit==3.7.1
